+++ src/apis/guides.ts
@@ -1,11 +1,9 @@
 import apiClient from './apiClient';
 import type {
   GetOfficialGuideProfileResponse,
-  GetGuideResumeResponse,
   GetPreGuideFeedbackResponse,
   GetPreGuideFeedbackRateResponse,
   GetPreGuideRateStatusResponse,
-  UpdateGuideResumeStepRequest,
   SubmitFeedbackRateFieldRequest,
 } from 'types/guides';
 
@@ -14,32 +12,6 @@ export const getGuideProfiles =
     const res = await apiClient.get('/guides');
     return res.data;
   };
-
-export const createGuideResume = async () => {
-  const res = await apiClient.post('/guides/resume');
-  return res.data;
-};
-
-export const submitGuideResume = async () => {
-  const res = await apiClient.post('/guides/resume/submit');
-  return res.data;
-};
-
-export const getGuideResume = async (): Promise<GetGuideResumeResponse> => {
-  const res = await apiClient.get('/guides/resume');
-  return res.data;
-};
-
-export const updateGuideResumeStep = async (
-  stepNumber: number,
-  payload: UpdateGuideResumeStepRequest
-) => {
-  const res = await apiClient.patch(
-    `/guides/resume/steps/${stepNumber}`,
-    payload
-  );
-  return res.data;
-};
 
 export const getPreGuideFeedbacks =
   async (): Promise<GetPreGuideFeedbackResponse> => {
