+++ src/types/guides.ts
@@ -9,21 +9,6 @@ export interface GuideProfile {
 
 export type GetOfficialGuideProfileResponse = GuideProfile[];
 
-export interface UpdateGuideResumeStepRequest {
-  answer: string;
-}
-
-export interface GuideResume {
-  name: string;
-  age: string;
-  nationality: string;
-  language: string;
-  expertise: string;
-  link: string;
-}
-
-export type GetGuideResumeResponse = GuideResume;
-
 export interface GuideFeedback {
   // now for pre guide
   feedbackId: number;
